pyproject.toml
[project]
name = "mt"
version = "0.1.0"
description = "mt is a simple desktop music player designed for large music collections."
authors = [
    { name = "pythoninthegrass", email = "4097471+pythoninthegrass@users.noreply.github.com" }
]
requires-python = ">=3.11,<3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "eliot-tree>=24.0.0",
    "eliot>=1.17.5",
    "fastapi>=0.117.1",
    "greenlet>=3.2.4",
    "jinja2>=3.1.6",
    "mutagen>=1.47.0",
    "pydantic-settings>=2.11.0",
    'pyobjc-framework-Cocoa; platform_system == "Darwin"',
    'pyobjc-framework-Quartz; sys_platform == "darwin"',
    "python-decouple>=3.8",
    "python-multipart>=0.0.20",
    "python-vlc>=3.0.21203",
    "pywebview>=6.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "tk>=0.1.0",
    "tkinterdnd2>=0.4.2",
    "uvicorn>=0.37.0",
    "watchdog>=6.0.0",
    "websockets>=15.0.1",
    "ziggy-pydust>=0.25.1",
    "ziglang==0.14.0",
]

[project.scripts]
repeater = "utils.repeater:main"

[project.optional-dependencies]
dev = [
    "pyclean>=3.1.0",
    "ruff>=0.9.6",
    "tkreload>=1.0.4",
    "ziggy-pydust>=0.25.1",
]

test = [
    "playwright>=1.55.0",
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
    "pytest-cov==6.0.0",
]

[build-system]
requires = ["hatchling", "ziggy-pydust==0.25.1"]
build-backend = "hatchling.build"

[tool.hatch.build]
script = "build.py"

[tool.hatch.build.targets.wheel]
packages = ["app", "core", "utils"]
artifacts = ["core/*.so"]
include = [
    "*.py",
    "static/",
    "themes.json",
]

# Configure ziggy-pydust to use our own build.zig
# It will still codegen pydust.build.zig
[tool.pydust]
self_managed = true
